Refactor admin bro role-permissions
