+++ package.json
@@ -162,7 +162,7 @@
     "test:categoryResolver": "NODE_ENV=test mocha ./test/pre-test-scripts.ts  ./src/resolvers/categoryResolver.test.ts",
     "test:givpower": "NODE_ENV=test mocha -b -t 30000 ./test/pre-test-scripts.ts ./src/repositories/powerBoostingRepository.test.ts  ./src/repositories/userPowerRepository.test.ts ./src/repositories/powerRoundRepository.test.ts ./src/repositories/userProjectPowerViewRepository.test.ts ./src/repositories/projectPowerViewRepository.test.ts ./src/resolvers/powerBoostingResolver.test.ts ./src/resolvers/userProjectPowerResolver.test.ts  ./src/adapters/givpowerSubgraph/givPowerSubgraphAdapter.test.ts ./src/repositories/projectRepository.test.ts ./src/resolvers/projectResolver.test.ts ./src/repositories/dbCronRepository.test.ts",
     "test:apiGive": "NODE_ENV=test mocha ./test/pre-test-scripts.ts  ./src/routers/apiGivRoutes.test.ts",
-    "test:adminBro": "NODE_ENV=test mocha ./test/pre-test-scripts.ts  ./src/server/adminBro.test.ts",
+    "test:adminBro": "NODE_ENV=test mocha ./test/pre-test-scripts.ts  src/server/adminBro/adminBro.test.ts",
     "test:poignArt": "NODE_ENV=test mocha ./test/pre-test-scripts.ts  ./src/services/poignArt/api.test.ts",
     "test:bootstrap": "NODE_ENV=test mocha ./test/pre-test-scripts.ts  ./src/server/bootstrap.test.ts",
     "test:utils": "NODE_ENV=test mocha ./src/utils/utils.test.ts",
@@ -176,7 +176,7 @@
     "prettify": "prettier --write '**/*.ts*'",
     "db:migrate:seedToken:run": "NODE_ENV=development ts-node ./node_modules/typeorm/cli -f ./src/seedToken-ormconfig.ts migration:run",
     "db:migrate:seedToken:revert": "NODE_ENV=development ts-node ./node_modules/typeorm/cli -f ./src/seedToken-ormconfig.ts migration:revert",
-    "build": "rm -rf ./build && tsc && mkdir ./build/config && mkdir ./build/src/server/components && cp -r ./src/server/components/* ./build/src/server/components/ && mkdir ./build/src/utils/locales && cp -r ./src/utils/locales/* ./build/src/utils/locales/ ",
+    "build": "rm -rf ./build && tsc && mkdir ./build/config && mkdir ./build/src/server/components && cp -r src/server/adminBro/components/* ./build/src/server/components/ && mkdir ./build/src/utils/locales && cp -r ./src/utils/locales/* ./build/src/utils/locales/ ",
     "dev": "NODE_ENV=development node ./build/src/index.js",
     "production": "NODE_ENV=production node ./build/src/index.js",
     "start:docker:server": "npm run db:migrate:run:production && npm run production",
