+++ src/.adminbro/.entry.js
@@ -1,3 +1,3 @@
 AdminBro.UserComponents = {}
-import Component1 from '../server/components/FilterListedComponent'
+import Component1 from '../server/adminBro/components/FilterListedComponent'
 AdminBro.UserComponents.Component1 = Component1
