+++ tsconfig.json
@@ -59,5 +59,7 @@
     "experimentalDecorators": true /* Enables experimental support for ES7 decorators. */,
     "emitDecoratorMetadata": true /* Enables experimental support for emitting type metadata for decorators. */
   },
-  "exclude": ["./build", "node_modules", "src/server/components"]
+  "exclude": ["./build", "node_modules",
+    "src/server/adminBro/components"
+  ]
 }
